Try to plot bottlenecks over time
But it's impossible to add a new measurement source and new report without strongly affecting API.
Raw results are not piped through. In more general case, there's no way to plug in more "raw" results.
The entire list of "supported" results is hardcoded within `FullCohortResult` and `EdibleResult`.
They're also very implicily dependent on the `infrastructure` implementation, which determines the shape of raw data on filesystem.